Cleanup user-profile types
